refactor(cockpit): move trend tests to owner module
